requirements.txt
fastapi
uvicorn
pdfplumber
faiss-cpu
python-dotenv
httpx
anthropic
python-multipart
numpy